+++ pages/collections/[id]/edit.tsx
@@ -20,151 +20,16 @@ import { GetStaticPaths, GetStaticProps } from 'next';
 import { Action, Pile } from '../../../types/types';
 import { getPiles } from '../../api';
 import { getPile } from '../../api/[id]';
+import EditPile from '../../../components/EditPile/EditPile';
 
-const { NEXT_PUBLIC_API_URL } = process.env
 interface Props {
   pile: Pile;
 }
 
-export default function EditPile({ pile }: Props) {
-  const exampleAction: Action = {
-    title: 'Scen på',
-    topic: 'light_mixer/code/run',
-    msg: 'scene_on',
-  };
-
-  const [advanced, setAdvanced] = useState<boolean>(false);
-  const [name, setName] = useState<string>(pile.name);
-  const [actions, setActions] = useState<Action[]>(pile.actions);
-
-  function createAction() {
-    setActions([
-      ...actions,
-      {
-        title: 'Scen på',
-        topic: 'light_mixer/code/run',
-        msg: 'scene_on',
-      },
-    ]);
-  }
-
-  function deleteAction(index: number) {
-    const newActions = [...actions];
-    if (index === 0) {
-      newActions.shift();
-      setActions(newActions);
-    } else {
-      newActions.splice(index, 1);
-      setActions(newActions);
-    }
-  }
-
-  function updateAction(value: string, index: number, field: 'title' | 'msg' | 'topic') {
-    const newActions = [...actions]; // copying the old datas array
-    newActions[index][field] = value; // replace e.target.value with whatever you want to change it to
-
-    setActions(newActions);
-  }
-
-  async function updatePile(id: string) {
-    await axios
-      .put(
-        `${NEXT_PUBLIC_API_URL}/${id}`,
-        {
-          name,
-          actions,
-        },
-        {
-          headers: {
-            'Content-Type': 'application/json',
-          },
-        }
-      )
-      .then((res) => {
-        showNotification({
-          title: 'Success',
-          message: `Updated collection \'${res.data.name}\'! Reload for changes to show.`,
-          icon: <Check size={18} />,
-          color: 'green',
-        });
-        Router.push('/collections');
-      })
-      .catch((err) => {
-        showNotification({
-          title: 'Error',
-          message: 'Something went wrong.',
-          icon: <X size={18} />,
-          color: 'red',
-        });
-        console.error(err);
-      });
-  }
-
+export default function Edit({ pile }: Props) {
   return (
-    <Container size="sm">
-      <Title order={1} mb={25}>
-        Editing {pile.name}
-      </Title>
-      <TextInput
-        label="Collection name"
-        placeholder="Gasque"
-        value={name}
-        onChange={(e) => setName(e.target.value)}
-      />
-      <Group mt="lg" spacing="xs" align="center" position="right">
-        <Text size="sm" color="dimmed">
-          Advanced mode
-        </Text>
-        <Checkbox checked={advanced} onClick={() => setAdvanced(!advanced)} size="xs" />
-      </Group>
-      <Divider my="lg" />
-      <Group mb="lg" position="apart">
-        <Title order={2} sx={{ fontSize: 18 }}>
-          Actions
-        </Title>
-        <ActionIcon onClick={() => createAction()} variant="filled" color="green">
-          <Plus size={18} />
-        </ActionIcon>
-      </Group>
-      <Group grow direction="column">
-        {actions.map((action, i) => (
-          <Card withBorder shadow="sm" key={i} mb="lg">
-            <Group position="right">
-              <ActionIcon onClick={() => deleteAction(i)} color="red" variant="filled">
-                <Trash size={18} />
-              </ActionIcon>
-            </Group>
-            <TextInput
-              label="Label"
-              onChange={(e) => updateAction(e.target.value, i, 'title')}
-              value={action.title}
-            />
-            {advanced && (
-              <TextInput
-                label="Topic"
-                onChange={(e) => updateAction(e.target.value, i, 'topic')}
-                value={action.topic}
-              />
-            )}
-            <TextInput
-              label={advanced ? 'Message' : 'Script name'}
-              onChange={(e) => updateAction(e.target.value, i, 'msg')}
-              value={action.msg}
-            />
-          </Card>
-        ))}
-      </Group>
-      <Button
-        fullWidth
-        color="green"
-        variant="filled"
-        leftIcon={<DeviceFloppy />}
-        onClick={() => updatePile(pile._id)}
-      >
-        Save
-      </Button>
-    </Container>
-  );
+    <EditPile pile={pile} />
+  ) 
 }
 
 export const getStaticProps: GetStaticProps<Props> = async (context) => {
